+++ package.json
@@ -69,7 +69,7 @@
     "headers-utils": "^1.2.0",
     "node-fetch": "^2.6.0",
     "node-match-path": "^0.4.4",
-    "node-request-interceptor": "^0.3.5",
+    "node-request-interceptor": "^0.5.0",
     "statuses": "^2.0.0",
     "yargs": "^15.4.1"
   },
